move parameter to mbql functions to own utils file
